:recycle: Allow colour to use integers and text formatting
Colour previously required constant names that were of a custom type.
This made conversion back to the primitive type difficult. Switching
to integers simplifies the process.

Inline formatting should occur when the colour is set. This was not
possible previously. Implemented the functional options pattern to
allow a variable number of arguments.