:recycle: Change components to models
Change components to individual models. Components are able to update
their own interface, act on messages (including keyboard commands) and
store state and configuration.

A toggle has been added which changes the component to have faint
content if it is not in focus. This applies to the summary and body
sections.