:recycle: Improve model state
The individual models need to track their focus state and update their
interface when this occurs. Messages and commands flow down from the
main model to component models.

If a component has focus, it should not allow focus to be reapplied as
this will cause a number of issues including loss of commands and extra
messages.