:children_crossing: Improve shortcuts
Shortcuts were inconsistent and confusing. Taking advantage of the
multiline status allowed the use of both alt and control to navigate
the interface.

The different components are accessed with the number keys which are
simpler to memorise.

Committing a change with control-enter was not possible with existing
terminals and had to be swapped to alt-enter.