:sparkles: Implement summary counter
Implement functional summary counter. The counter colour changes based
on the length of the summary. Follows the recommendations from popular
git linters.

These recommendations are:
- < 5 is yellow.
- > 5 and <= 40 is green.
- > 40 and <= 50 is yellow.
- > 50 is red.

Subject length has a limit of 72 characters. Until the emoji component
is implemented this will fail to be accurate as it will need to take
into account if an emoji was added to the subject line.