:recycle: Improve model state tracking
The model state tracks the focused model rather than the component which
is active. Having a more granular state allows for components within
models to act differently depending on which part is active.

This improvement will allow for the summary and emoji components to be
aware of their focus state as well as the global state to be tracked at
a component level. Tracking the component at the global level is
necessary to have interface changes in multiple models at the same time.

A display value was added to define different layouts depending on the
component that is active. This can be applied to multiple models at the
same time.