:bug: Improve emoji padding
Emoji alignment is difficult due to the nature of them having multiple
widths. They can be double character, ambiguous and narrow which makes
the layout inconsistent with other elements especially when there are
borders.

Detecting the rune width allows for some of these cases to be handled
correctly by using spaces as padding.