:bug: Align lines with emojis correctly
Alignment issues occur when some lines have certain emojis in them. This
is caused by incorrectly calculating the width of a string.

An upstream dependency of Bubbletea is providing the wrong result and a
pull request has been created to add a function to the styling package
(Lipgloss) to fix this issue.