:sparkles: Build Git commit command
Build the command that will be executed by Git. Some refactoring was
necessary for the commit struct to hold both the commit and the original
config.

Templating is used to build the commit message as this provides a more
flexible method for updating. In the long term this will be replaced by
native Git commits within the application.

The keyboard shortcut to accept a commit needs to validate that the
fields necessary for a successful commit exist. If there is no summary
the action will not be accepted.