:recycle: Convert header component into package
Convert header component into a package to abstract away the main model.
This was the final component that required refactoring and meant unused
code could be removed.

By splitting up the header component it allowed the filter to be
converted into a Bubbletea model. The filter list is a generic component
and has no knowledge of the structure of the items it will display.

The paginator was improved to be more reusable and has support for
horizonal pagination which is currently not used.

The emoji package was simplified as it did not need to provide
methods used by both the list and the fuzzy matching. Casting with the
required methods is done within the header package.