:sparkles: Implemented author switching
Add capability to switch the author profile. Only the default system or
repository author is detected but a configuration component will be
added to allow for this to be set from a config file.

To enable this functionality some refactoring was required with the
filter list to support variable height.