:bug: Fix body message format
The value of the body message was not output as shown. Line wrapping was
removed and each line was based on the terminating new line character.

By using the view output rather than the value, this provided the
correct layout as this was the representation as it was shown on the
terminal. This required stripping of ANSI and trailing spaces as these
were part of the view.