:package: Add theme package
Add theme package which uses the BubbleTint library.

Colours are used in many components so this was best implemented using
a singleton. Passing the struct through many function signatures would
have added complexity.

Tint cycling will be used for a simple implementation, however this
will be replaced at a later date with a configuration option.