:children_crossing: Add ttyd compatibility mode
The terminal ttyd has all emojis regardless of string width as 1
character. This means that alignment is consistent and padding is
required for all emojis.

This is not typical with how other terminals handle emojis and most
terminals differentiate between emojis of string with 1 or 2. Detection
is required to determine the correct padding.