:dizzy: Update demo and tape
Update the demo to render at a larger font size. This reduces the
blurriness of the font.

Other minor changes to the demo loop included typing the command as
well as showing some of the other emoji pages.

The demo tape was updated to include many of the manual steps that had
to be prepared before recording. This has reduced the prerequisite
steps.

By increasing the dimensions of the demo, the html within the markdown
could be removed as left justification is no longer needed.