:white_check_mark: Refactor repository user tests
Repository user tests were not testing failing or complex cases. Using
the go-git fixtures testing suite was limiting.

By using interfaces, the package could be tested more extensively while
ignoring the responsibilities of the go-git package. This has created
more robust checks that include error states.

Some refactoring was required for the repository that changed the user
field in the struct to a method.