:white_check_mark: Refactor repository branch tests
Repository branch tests were refactored to allow for testing errors and
remove the reliance on go-git fixtures.

This also applies the improvement to use an interface for the
repository. This matches the changes that were applied to the other
repository functions.

Significant refactoring was applied to the branch method to simplify
and improve clarity. The private methods were changed to functions and
if many options were required a struct was passed.