:bug: Fix panic when local branch has no commits
When trying to determine the remote for a branch there is an edge case
that causes a panic.

If the local branch exists but does not have any commits, the map does
does not have a key for this branch causing a nil pointer exception
when looking up the remote.

This seems to be a bug in go-git as all branches should be returned
with empty structs.