:recycle: Move find emoji functions
The emoji find functions were inappropriate as part of the commit
package. This was more suited to being part of the emoji package.

This refactor made it necessary to return a null emoji if not found.
Previously this was an empty emoji which didn't convey a negative
result. Instead of using a null pointer for the null result, this was
changed to the type NullEmoji which contains a field with the result
status.

HasEmoji was also renamed as it was confusing and did not make it clear
exactly which field it was returning.