:recycle: Refactor commit package
The commit package transforms the repository descriptions into its own
structure. This did not provide any benefits and could be removed.

By adding a describe method to the repository package, a single call
can provide all the users, remotes, head and branch details.

For amends, the message transformation still needs to be done in the
commit package so these fields needed to be added to the config struct.