:recycle: Refactor repository commit
Refactor repository commit to be more modular and testable. This
involved splitting the shell runner into a separate package.

The output function was no longer needed as stdout could be used as a
writer for the runner. This required some extra padding for the
message.

Tests were added to verify the results.