:children_crossing: Consolidate apply and dry run
There was inconsistence in the use of both apply and dry run as
options. The Git command uses dry run so all apply options have been
renamed.

Some logic needed to be inverted however the boolean is now the same in
all places.

The developer apply option has been removed as it was redundant.