:recycle: Refactor commit apply
The commit apply now accepts a request object rather than requiring the
field to be set in the struct.

Transforms did not need to be methods as they only required a few
parameters.

Combining these two refactors improved testing and made development
simpler as the commit package is easier to understand.