:recycle: Improve filterlist with dynamic width
The filterlist had a set width which was limting its use in other
areas. The character limit was also fixed. Refactored internally to
make this more dynamic.

These changes exposed a bug with the paginator dots that caused them to
not be properly aligned with the right border. A space is meant to be
between the dots and the border and this did not occur for the author
list.

Another bug was found with the text input prompt in the filterlist. The
width of the prompt was not taken into account and caused an incorrect
width for the entire input.

As a side effect of these changes, the info unit tests have failing
golden files which has been corrected.