:recycle: Split shortcuts into a model
Split UI status component into a status and shortcut model.

This change will allow for customisable shortcuts without setting any
defaults. A global shortcuts function now provides the values for the
UI.

There was also a bug with modifier labels that contained whitespace.
Labels now have the whitespace trimmed before checking if the plus sign
should be added.

With shortcuts now being customisable, a bug with showing modifiers
when there are empty keys has been resolved. This has caused an update
to a golden file which provides a different result.