:white_check_mark: Add Cobra command tests
Add tests for the Cobra command entrypoint. Only flags are necessary to
check as there are no arguments.

The RunE and Execute functions are not testable as they require a real
repository and cannot be mocked.

The version template was changed to send output to the same writer that
is configured for Cobra. This allows the version details to be captured
and tested.