:recycle: Refactor UI to allow for testing
Refactor main UI component to allow for testing. By adding a Start
function, this allowed for New to only provide a model.

The Start function initialises a new Bubble Tea program which is
blocking so testing this function is not possible.

The response was made a public field to allow for verification of the
commit values returned on apply.