:recycle: Refactor shortcut naming
The shortcut naming scheme was confusing and ended up causing
repetition of the shortcut term.

This change introduces the concept of key bindings.

Shortcut -> KeyBinding
ShortcutSet -> KeySet
Shortcuts -> KeyBindings + Modifiers

The global and help shortcuts now return a single value instead of
multiple values. This reduces the code to assign these to the shortcut
model.