:recycle: Improve sign-off toggling
Sign-off toggling required some additional changes after the child
models were set. The change to allow setting the message to nil allowed
both the footer and body to update at the same time.

Previously this occured immediately for the footer and the next cycle
was the body. This created a visual delay to the animation.