:white_check_mark: Improve root command tests
Testing the Cobra root command was difficult as the application was not
able to be mocked with the current architecture.

Refactoring the root command to accept an app struct which contained
interfaces for the commit, UI and logger has allowed for improved
testing.

The configuration, running the UI and applying the commit are now done
in different Cobra steps. These steps have been split into functions to
help keep the root command functions simple.

Creating a new commit and UI model now only returns the struct. A new
method has been created to configure these components. This is to
improve testability with mocking the main loop.

Some small changes to use pointers instead of values was required
depending on the struct size.