:recycle: Refactor commit configure
To create tests for the commit configure method, refactoring was
required. The dependencies were changed to interfaces to allow mocking.

A describer is used to describe an existing repository. If one has not
been set, the default is used. This allows for the describe method
which is returned from creating a new repository to be replaced with a
mock.