:sparkles: Add config loading when creating a new commit
Config loading is now implemented as part of creating a new commit.
This is provided as part of the state to all the UI components allowing
them to modify their behaviour based on options.

The state struct was refactored to add this config but also changed
options to be its own field. This improved the consistency. The config
file location was added as part of the options.

To help keep the configure function simple, the repository and config
initialisation was moved into their own private methods.

Functions and structs related to state were moved to their own file to
help keep the commit file manageable.

Some UI options were updated to use the new state structure. This was
to preserve existing functionality and not to add new features.