:recycle: Remove global variables for themes
The theme package was using global variables which is not a recommended
pattern. This made testing difficult as reseting the global variable
was required for each test.

The theme package is used in many packages related to the UI. This
required significant updates to function signatures. Providing the
entire state instead of just the config allows future customisation of
most of the UI without a large refactor.

The KeySetConfig was added as the function signature was becoming too
large after adding the state.