:memo: Add table of contents
Add table of contents using badges to keep a compact layout.

Some of the emojis for the headings had to be replaced with alternative
versions to allow for linking to headings. This may not display
correctly in some editors.

Added a link to the heading sections to jump back to the top of the
page.