🐳 Containerise demo video recording
Use containers to simplify the process of recording the demo video.
This also provides consistentency as changes on the host (such as
package updates) no longer impact the result.

It was necessary to set `LetterSpacing` to `0` as the containerised
output had wider letter spacing than the original recording under
macOS. This can be attributed to VHS release [`0.3.0`][1] which modified
the spacing without explanation.

A task has been created to trigger the recording.

[1]: https://github.com/charmbracelet/vhs/commit/026c21d2e1f7aa0bedf8409915361bc1b84fba18