:memo: Update table of contents
The width of the table of contents badges was causing overflow on some
browsers.

Reducing the badges to only a subset (the most important) solves this
problem and also cuts down on the visual clutter.

Move the table of contents above the animated image to improve layout
and usability.

Some of the section emojis could be restored to modern emojis as they
were no longer used in reference links.