:lipstick: Fix width of shortcut blocks
The left and right shortcut blocks were joined together with fixed
sizes. This meant the width of the left block combined with the width
of the right prevented spare space being allocated to either block.

Using hidden borders also reduced the usable width as the borders are
not rendered but are still part of the calculated width.