:sparkles: Add toggling between new and amend commits
Add feature to toggle between new and amend commits. The state is
preserved between the two different types which helps when deciding
which type of commit to apply.

The existing flag can still be used to set the initial state.

Switching state resets the cursor to the beginning of each input.

Further code changes are required to apply the commit based on the
current state.