:sparkles: Add loading and saving of cancelled commits
Add capability to load commits that were previously cancelled. Both new
and amend commits are supported and the saved state will be loaded into
either the new or amend save bank. This allows for toggling between the
two types without losing contents.

To aid in unit testing new commits, a fixed date was required. This
field was added to the main model which can be passed down to component
models.