:building_construction: Refactor commit architecture
Commit package was confusing with its use of storing functions and
interfaces within structs.

Consolidated Repoer to include applying the commit. There was no
benefit to having this separate and increased complexity for use and
testing.

Config loader and shell runner also had the same issues and were
changed to be stored as interfaces instead of functions.