:bug: Remove global variable used by help
A global variable contained the help message. When running tests in
parallel that modified the viewport this was causing contention issues.

The help text was moved to the commit package where other embedded
variables exist and is part of the state.

The help viewport now has a way to set the content which allows for the
text to be updated instead of writing to the global variable. Unit
tests results have changed for the default help state which is now
empty.