:test_tube: Fix repository branch tests
The repository branch tests were failing when run in a Docker
container.

When run locally, the global Git user can be used for signing commits.
However, in a container this will default to an empty user. Creating
tags requires a commit and all commits need to be attributed to an
author.

This change adds a signer when creating tags to resolve this problem.