:memo: Add comparison table to documentation
Add a comparison table between Committed and Gitmoji CLI. This
comparison highlights the missing features for Committed and will act as
a roadmap for future functionality.

As part of this change, the checkmarks and crosses were updated in other
parts of the documnetation to match the comparison table. These new
images are external instead of emojis and provide a more stylish design.