:construction: Add hook install function
Add function to create the hook. This checks for an existing file and if
it does not exist creates a file containing the hook script.

If the file does exist, the first line is scanned for a marker to
determine who created the file and overwrites it if it was from
Committed.