:sparkles: Add hook support
Git supports hooks that can run before the editor is displayed. A
prepare commit message hook was added to apply a commit using the
default Git commit command.

An install and uninstall flag was added and checks for a signature. This
signature prevents an unmanaged hook from being updated or deleted.

By default Git adds comment blocks to the message file. These comments
were displayed when applying a commit. A change has been made to the
output to remove these comments to reflect the actual commit message.

The hook will not be used for commits that are classified as merge,
squash or template.

There are a number of limitations in the existing implementation:

- Changing the author is not supported. The hook only allows updating
  of the message and no other fields.

- The default editor will be displayed after the commit is accepted.
  This is how prepare commit message hook works as it has been designed
  to work before the editor is started to modify the default message.