:bug: Fix hook location detection
Hook location detection had flawed logic and was only checking for a
global hook. If the global hook was not detected it was considered an
error and aborted with a failure.

Both global and repository hooks now do not consider failure if they
cannot find a hook. This allows for the true error message to be
reported if neither location can be detected.

The only case for a repository failure is if the hook commands were run
while not in a Git repository and the global hook location is unset.