:recycle: Simplify commit message output
The commit message text that is output on applying a commit is generated
by combining the emoji, summary, body and footer. Each component has a
customisable style however the output for all was the same. Removing
these styles and combine with a single style for the entire message
reduced the complexity significantly.

The final view was rendered using vertical dividers. This was causing
line spacing issues when components were removed. Changing this to a
slice allows for joining after all the elements have been added. This
can then be styled afterwards.

Removal of comments was problematic and has been removed. This will be
reimplemented at a later date when editor mode is added.