:sparkles: Add Git editor support
Add functionality to act as a Git editor. This is similar to hooks
however this resolves the issue of duplicate message editors.

By taking over the editing capabilities, there are some complexities
with merge and rebase where Git commands can be provided. Comment blocks
also needed to be trimmed to prevent reflow converting the new lines
into commands (which were not understood and failed).

There are likely a number of edge cases not handled (or considered) and
this feature should be used with caution and is experimental.

Implementation allowed for merging the hooks and editor functionality
into reading a file as they were very similar.