:bug: Fix reflow occuring unexpectedly in body
Reflow when entering text in the body was inconsistent. The following
test case shows the issue (has padding of 1 on both sides):

+------------+
| 1234567890 |
| 1 2 3 4 5  |
| 6          |
+------------+

The 10th character on the second line should be 6 however as there is a
space it is immediately reflowed to the next line. The cursor never being
allowed in the 11th character position creates this edge case.

This bug has further details in the GitHub issue:
https://github.com/charmbracelet/bubbles/issues/333

The textarea width is set to 72 characters however any line with a space
contained in it will reflow at 71 characters. To workaround this issue
reflow has been set to 71 characters.

Test cases have been added to detect this problem.