:bug: Improve format of saved config file
When the options are saved, they are encoded in YAML flow style. This
makes reading and modification difficult.

Flow mode was previously necessary as inline tests of YAML had
whitespace issues.

Refactoring the unit tests to handle whitespace has allowed flow
encoding to be removed and this resolves the issue with the config
file.