:bug: Fix saved config overwriting authors
When saving the configuration the authors field is not preserved. This
is because authors are not editable in the options and therefore are
not part of the new config struct which replaces the existing one.

Merging the old and new config structs together allows the existing
authors to be preserved. The merging library used does not merge zero
value fields.