:construction_worker: Fix Docker image builds
Building a release Docker image is different to building a functional
image locally. GoReleaser compiles the binary and makes it available in
the current directory. The binary should be copied into the image and
not built again.

For the release process, a variant Docker image was created that copies
the binary only and sets up a default Git user and empty repository.