:bug: Fix merging config changes
When setting config options changed within the UI, zero values were not
handled correctly. It was not possible to turn a value from true to
false or any other zero value.

There was no need to merge the existing config with the options set
from the configuration. All values that can be set within the config
struct are set either with the values within the options panels or from
the existing state.

It was necessary to refactor the function to allow for easier testing.
This meant providing a result and accepting parameters for the input.