⬆️ Upgrade dependencies
Upgrade third-party libraries to their latest stable releases to
incorporate security patches, bug fix and performance improvements.
This routine maintenance helps keep the codebase secure and efficient.

As part of this upgrade, some regressions were identified and
corrected.

Using page up and page down in the filter list was being sent twice. A
bug with the list component was fixed upstream and the existing code
attempted to send this twice however one of the methods was a noop.
This was corrected by removing the unnecessary duplication.

A list with empty items had the wording changing. This required the
golden files to be updated.

All dependencies were determining string width using `go-runewidth`
which has been problematic. This has been replaced with `uniseg` which
required the `StringWidth()` function to be swapped over to be
consistent. This also caused some incorrectly rendered golden files to
be fixed.

This upgrade will cause the terminal package to activate the grapheme
cluster width override map. This will allow the rendering of emoji
widths to be cosnsistent to previous versions.

The upgrade to Bubbletea has caused a minor change in the quit message
type which has become public.