♻️ Update VHS commands
VHS is used to record the demo video that is part of the README.

As part of the scripting process to generate the video, some extra
capabilities needed to be added to a forked VHS.

The keyboard presses were:
- Shift+Tab
- Alt+Enter

Both these commands have now been added to the official VHS release.
However, there were minor changes in how the commands were written
which is reflected in this change.