🐛 Hide end of buffer character
The end of buffer character is currently enabled however it is not
shown except for when the content does not fill the viewport for the
placeholder.

This functionality is being changed in an upcoming release of the
textarea component and will make this character appear by default.

This change is preemptive to reduce the number of files changed as part
of the dependency upgrade.