✨ Add terminal package
As part of the upcoming dependency upgrade, there will be changes to
the supported Unicode version. This change will cause some emojis
(grapheme clusters) to report different string widths than previously.
This will impact the layout especially when there are borders.

The terminal package will have the role of handling compatibility
issues between the terminal and packages used to determine string
width.

The first functionality added to this package is to allow the
overriding of the width of specific grapheme clusters (a selection
using variant select 16). This will provide compatibility for any
terminal that does not currently support Unicode 14 or above.

For overriding of grapheme cluster width, a fork of `uniseg` was
necessary to allow for an replacement map.

While the width calculcation changes are applied when the UI is
configured, no packages are currently using the `uniseg` functions so
none of the overrides apply. Once the dependency upgrade is completed,
this will provide the same results as the previous string width
functions.