⬆️ Upgrade dependencies
Upgrade dependencies to latest releases. As part of this upgrade there
were two packages worth mentioning.

The `bubbles` package was previously using a modified fork that added
support for [textarea multiline placeholders][1]. This change has now
been merged upstream and there is no longer a need to rely on the
custom version.

The `coloredcobra` package which provides the colours for the CLI help
text was impacted by the upgrade of the `color` package. A [fix][2] has
been applied that currently only exists in a forked version. This is
necessary until the upstream merges the pull reques to fix this
problem.

[1]: https://github.com/charmbracelet/bubbles/pull/302
[2]: https://github.com/ivanpirog/coloredcobra/pull/5