♻️ Replace reflow package
The `reflow` package is used to word wrap the commit description. This
is necessary to make the text sent to Git match the editor view.
Unicode support was not implemented correctly by this package and would
cause the results to be inconsistent with what was displayed. The
replacement package `ansi` resolves this problem.