🐳 Upgrade Docker images
The Docker images were using older base images that needed to be
upgraded. Pinning of the packages was problematic as Alpine does not
keep older packages. By relaxing the restrictions to ignore minor and
patch package changes should prevent future package installation
failures.

It was also necessary to fix the case of the `as` directive as Docker
was emitting warnings on build.