🐛 Restore snapshot for all unapplied commits
When a commit fails to be applied, the message is saved in a snapshot
state file. However, there are other conditions where the snapshot is
not used:

- Not staging any files. Only option is quit the application.

- Quiting without committing.

These conditions occur when the user presses `ctrl-c` after they have
started to write a message.

We should always try to preserve any existing message so that it can be
reused for the next session.