🧪 Fix UI tests default state
The UI tests set up a default state that was used by all the test
cases. The commit type was set to amend which is not the default
commit when normally using the application. Switching this allowed
tests to no longer always set the amend field when checking the
resulting state.

There was a large impact on the golden files as they showed the commit
type in the header. Updates were required to match them to the expected
result.

A problem was identified with the placeholder hash not being set for
new commits as part of the test state. This default was also included
to provide consistency and some golden files needed to be updated as
the hash was empty.