🐛 Fix snapshot restore
When restoring a snapshot, the commit type was not taken into account.
It was not useful to restore the snapshot when the commit type of the
snapshot differs to the requested state.

This means we should not restore an amend commit snapshot into a
standard commit or a standard commit snapshot into an amend commit.

This logic now makes the load option unnecessary and can be deprecated.

Some unit tests needed to be updated as they did not take into account
the typical snapshot restore cases.